Add browser service capabilities
and rework the Browsers class to incorporate inner classes for a namespace orientated structure.

This uses a BrowserServiceConverter which can crete a DriverService (in the future, this might actually convert a BrowserServiceContext, similar to the BrowserContext).
However, the BrowserServiceSupplier is responsible to call the converter and thus can decide which browser should use a service.

Lastly, the BasicBrowserConverter has been rewritten to have dedicated factories, capable of accepting both a context with and without a service.
This is to anticipate the optional service handling done by the BasicBrowserConverter.